requirements.txt
python-telegram-bot>=20.0,<21.0
solana>=0.25.0
solders
python-dotenv 
telegram